pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "werewolf-bench"
version = "0.1.0"
description = "Werewolf assessment harness (migrated from tau-bench example)."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "earthshaker>=0.1.12",
    "pydantic-settings>=2.11.0",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]
